# improve performance of `strawberry.lazy()` with relative imports