--- a/RELEASE.md
+++ b/RELEASE.md
@@ -1,0 +1,3 @@
+Release type: patch
+
+This release includes a performance improvement to `strawberry.lazy()` to allow relative module imports to be resolved faster.