<!--- Provide a general summary of your changes in the title above. -->

<!--- This template is entirely optional and can be removed, but is here to help both you and us. -->
<!--- Anything on lines wrapped in comments like these will not show up in the final text. -->

## Description

<!--- Describe your changes in detail here. -->

I've been profiling a few tests in my company's codebase and realized that some tests that actually only perform about 1s of business logic take ~12s to run. A large part of that time (~8-9s) is spent resolving forward refs in `strawberry.lazy()`. The largest culprit there was the use of `inspect.stack()` which seems to be incredibly slow.

This PR updates the method of retrieving the target stack frame. It seems that the stack frames are stored as a linkedlist anyways, so retrieving the active frame and then working backwards is significantly faster (likely because `.stack()` collects more information for every frame that is unused by strawberry). This one line change took my test down from 12s to 2.5s.

## Types of Changes

<!--- What types of changes does your pull request introduce? Put an `x` in all the boxes that apply. -->
- [ ] Core
- [ ] Bugfix
- [ ] New feature
- [X] Enhancement/optimization
- [ ] Documentation

## Issues Fixed or Closed by This PR

N/a

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [X] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
- [ ] I have added tests to cover my changes.
- [X] I have tested the changes and verified that they work and don't break anything (as well as I can manage).