@patrick91 2023-07-06
as a not for myself, I think we should add some benchmark for this :)